+++ src/main/java/szewek/mcflux/blocks/BlockEnergyMachine.java
@@ -30,7 +30,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class BlockEnergyMachine extends BlockContainer {
-	public static final AxisAlignedBB DEF_AABB = new AxisAlignedBB(0.25, 0.25, 0.25, 0.75, 0.75, 0.75);
+	private static final AxisAlignedBB DEF_AABB = new AxisAlignedBB(0.25, 0.25, 0.25, 0.75, 0.75, 0.75);
 	public static final PropertyEnum<Variant> VARIANT = PropertyEnum.create("variant", Variant.class);
 	public static final PropertyInteger
 			UP = PropertyInteger.create("up", 0, 2),