+++ src/main/java/szewek/mcflux/util/InjectRegistry.java
@@ -1,9 +1,13 @@
 package szewek.mcflux.util;
 
-import static java.lang.annotation.ElementType.TYPE;
+import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
 @Target(TYPE)
+@Retention(RUNTIME)
 public @interface InjectRegistry {
 	boolean included() default false;
 	String[] detectMods();